+++ lib/adyen_checkout.dart
@@ -20,6 +20,8 @@ export 'src/common/model/payment_method_configurations/google_pay/shipping_addre
 export 'src/common/model/payment_method_configurations/stored_payment_method_configuration.dart';
 export 'src/common/model/payment_result.dart';
 export 'src/common/model/result_code.dart';
+export 'src/components/action_handling/model/action_component_configuration.dart';
+export 'src/components/action_handling/model/action_result.dart';
 export 'src/components/apple_pay/adyen_apple_pay_component.dart';
 export 'src/components/apple_pay/model/apple_pay_button_style.dart';
 export 'src/components/apple_pay/model/apple_pay_button_theme.dart';