+++ lib/src/components/action_handling/model/action_result.dart
@@ -1,0 +1,13 @@
+sealed class ActionResult {}
+
+class ActionSuccess extends ActionResult {
+  final Map<String, dynamic> data;
+
+  ActionSuccess(this.data);
+}
+
+class ActionError extends ActionResult {
+  final String errorMessage;
+
+  ActionError(this.errorMessage);
+}