+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## 0.1.1
 
-* Made amount optional. For Google Pay and Apple Pay in combination with the advanced flow, the
-  amount is still required.
+* Made amount optional. For Apple Pay in combination with the advanced flow, the amount is still
+  required.
+* Added standalone action handling.
 
 ## 0.1.0